AI Games/Chess AI/requirements.txt
pygame==2.5.2
python-chess==1.10.0 